# shared-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Veresiye Platform API
  version: 0.1.0
  description: |
    The single contract app-pos, app-mobile and the backend all speak (see
    veresiye-platform-tasarim.md §4.1). Generated clients / a Prism mock come from
    this file. Source of truth for endpoints/bodies: docs/api-endpoints.md; for the
    data shapes: docs/db-schema.md.

    Conventions:
      - Money: every *_minor is an integer in minor units (kuruş). 50 TL = 5000. No floats.
      - Scaling: orderBody item quantity and tax_percent are ×1000 (1000 = 1 unit / 10%).
      - seller_id / buyer are taken from the bearer token on scoped endpoints, not the body.
      - Field names are snake_case on the wire; Kotlin domain is camelCase (mapper converts).

    Tags:
      - "active"  = needed now (both clients' real mock behaviour). Fully modelled.
      - "future"  = planned but not wired yet (schema/contract placeholder; code skeleton
                    lands in Room but stays unconnected — see plan Aşama 3).
servers:
  - url: http://127.0.0.1:4010
    description: Prism mock (npx @stoplight/prism-cli mock openapi.yaml)
  - url: https://api.veresiye.example
    description: Backend (later phase)
...
security:
  - bearerAuth: []
paths:
...
  /approvals:
    post:
      tags: [approvals]
      summary: Send for approval (one of the three lines)
      description: |
        CLAIMED target -> a PENDING Approval (app-push card). UNCLAIMED target ->
        the entry is written immediately (SMS-OTP mock) and a Transaction returned.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalCreate' }
      responses:
        '201':
          description: Pending (CLAIMED target)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '200':
          description: Written immediately (UNCLAIMED target)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
    get:
      tags: [approvals]
      summary: Approvals addressed to the signed-in target
      description: >
        Defaults to PENDING (the inbox). Clients POLL this and sync their local table to
        the answer, so the response is authoritative: a row that is absent has been decided
        elsewhere and must be dropped locally.
      parameters:
        - name: status
          in: query
          required: false
          description: Defaults to PENDING. `ALL` returns the decided history too.
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, ALL]
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100, minimum: 1, maximum: 500 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Approval' }
              # Two seed approvals (app-mobile): p1 waits on u1, p2 on u_owner.
              example:
                - { approval_id: p1, initiator_user_id: u_owner, initiator_role: SELLER, target_user_id: u1, seller_id: u_owner, shop_name: Ahmet Bakkal, customer_id: c1, amount_minor: 5000, type: DEBT, description: 'Ekmek, süt', channel: APP_PUSH, status: PENDING, requested_at: '2026-07-25T07:05:00Z', updated_at: '2026-07-25T07:05:00Z' }
                - { approval_id: p2, initiator_user_id: u_market, initiator_role: SELLER, target_user_id: u_owner, seller_id: u_market, shop_name: Ayşe Market, customer_id: o1, amount_minor: 7500, type: DEBT, description: Temizlik malzemesi, channel: APP_PUSH, status: PENDING, requested_at: '2026-07-25T08:20:00Z', updated_at: '2026-07-25T08:20:00Z' }
  /approvals/{id}/approve:
    post:
      tags: [approvals]
      summary: Approve -> writes the ledger entry
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Written
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
  /approvals/{id}/reject:
    post:
      tags: [approvals]
      summary: Reject -> no write
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Rejected }

  # ----------------------------------------------------------------- future (placeholders)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    TransactionType:
      type: string
      enum: [DEBT, PAYMENT]
...
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    InitiatorRole:
      type: string
      enum: [BUYER, SELLER]
    Channel:
      type: string
      enum: [APP_PUSH, SMS_OTP]

    # --- auth ---
...
    OrderItem:
      type: object
      required: [name, price, quantity, tax_percent]
      properties:
        name: { type: string }
        price:
          type: integer
          format: int64
          description: 'per-unit, minor units (kuruş)'
        quantity:
          type: integer
          format: int64
          description: 'scaled ×1000 (1000 = 1 unit)'
        tax_percent:
          type: integer
          format: int64
          description: 'scaled ×1000 (1000 = 10%)'
        section_no: { type: integer, default: 1 }
        status: { type: integer, default: 1 }
        type: { type: integer, default: 0 }
        item_limit: { type: integer, format: int64, default: 0 }
    OrderBody:
      type: object
      required: [basket_id, items]
      properties:
        basket_id: { type: string, format: uuid }
        create_invoice: { type: boolean, default: false }
        document_type: { type: integer, default: 0 }
        is_void: { type: boolean, default: false }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }

    # --- transaction ---
    Transaction:
      type: object
      required: [transaction_id, seller_id, customer_id, amount_minor, type, description, settled_via_pgw, created_at]
      # Mirrors FakeRepository seed t1 (Ahmet's first debt, 50,00 TL).
      example:
        transaction_id: t1
        seller_id: u_owner
        customer_id: c1
        amount_minor: 5000
        type: DEBT
        description: Ekmek, süt
        basket_id: null
        basket: null
        settled_via_pgw: false
        receipt_no: null
        created_at: '2026-07-20T06:15:00Z'
      properties:
        transaction_id: { type: string, format: uuid }
        seller_id: { type: string }
        customer_id: { type: string }
        amount_minor: { type: integer, format: int64, minimum: 0 }
        type: { $ref: '#/components/schemas/TransactionType' }
        description: { type: string }
        basket_id: { type: string, format: uuid, nullable: true }
        # The basket itself, not only its id. Nothing serves a basket on its own, so an
        # id without the items beside it is a reference the client cannot resolve.
        basket:
          allOf: [{ $ref: '#/components/schemas/OrderBody' }]
          nullable: true
        settled_via_pgw: { type: boolean, default: false }
        receipt_no: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    Approval:
      type: object
      # Reconciled with app-mobile's real PendingApproval (shop_name, requested_at)
      # PLUS the forward three-line fields (initiator_*/target_user_id/channel). The
      # mock derives the three-line fields today (see plan Tur 21); full three-line
      # behaviour + app-pos's approvals UI land in a separate turn.
      required: [approval_id, initiator_user_id, initiator_role, target_user_id, seller_id, shop_name, customer_id, amount_minor, type, channel, status, requested_at, updated_at]
      # Line 2: seller-POS -> buyer-mobile, awaiting the buyer's approval (seed p1).
      example:
        approval_id: p1
        initiator_user_id: u_owner
        initiator_role: SELLER
        target_user_id: u1
        seller_id: u_owner
        shop_name: Ahmet Bakkal
        customer_id: c1
        amount_minor: 5000
        type: DEBT
        description: Ekmek, süt
        channel: APP_PUSH
        status: PENDING
        requested_at: '2026-07-25T07:05:00Z'
        updated_at: '2026-07-25T07:05:00Z'
      properties:
        approval_id: { type: string }
        initiator_user_id: { type: string }
        initiator_role: { $ref: '#/components/schemas/InitiatorRole' }
        target_user_id: { type: string }
        seller_id: { type: string }
        shop_name: { type: string, description: Denormalised for the buyer's card }
        customer_id: { type: string }
        amount_minor: { type: integer, format: int64 }
        type: { $ref: '#/components/schemas/TransactionType' }
        description: { type: string, nullable: true }
        channel: { $ref: '#/components/schemas/Channel' }
        status: { $ref: '#/components/schemas/ApprovalStatus' }
        requested_at: { type: string, format: date-time, description: When it was ASKED; never changes }
        updated_at: { type: string, format: date-time, description: When it last CHANGED; moves on every status transition }
    ApprovalCreate:
      type: object
      required: [seller_id, customer_id, amount_minor, type, initiator_role, target_user_id]
      properties:
        seller_id: { type: string }
        customer_id: { type: string }
        amount_minor: { type: integer, format: int64 }
        type: { $ref: '#/components/schemas/TransactionType' }
        description: { type: string, nullable: true }
        initiator_role: { $ref: '#/components/schemas/InitiatorRole' }
        target_user_id: { type: string }
        # Which device raised this. Only PHONE leaves gateway work behind: a POS is
        # already in front of the gateway and hands its own intent over.
        origin: { type: string, enum: [POS, PHONE], default: PHONE }
        # The basket the gateway handed over, when the sale came from one. Carried on the
        # REQUEST because the handoff is over by the time anyone approves.
        basket:
          allOf: [{ $ref: '#/components/schemas/OrderBody' }]
          nullable: true

    # --- future ---
...